Move naming logic to separate file
