+++ backend/src/api/report/mod.rs
@@ -4,6 +4,7 @@ use crate::{AppState, api::middleware::authentication::RouteAuthorization, domai
 
 mod files;
 mod handlers;
+mod naming;
 mod structs;
 
 /// Default date time format for reports
